operator/bluegreen-agent/requirements.txt
fastapi>=0.115.0
uvicorn==0.29.0
asyncpg==0.29.0
python-dotenv==1.0.1
setuptools>=78.1.1
starlette>=0.40.0